torch: implement passing values between pytorch and our code
This commit adds a new crate `torch` that bridges our code with PyTorch.
We use PyO3 crate to get PyTorch values from Python to our code and back
to Python.

The create uses three main structs: `PyVal` which is an enum that wraps a
value with its type (subject to change in the future),
`PyTerm` which is a version of `base::Term` that we can pass into/from Python
(passing `base::Term` directly does not work as the object cannot be
generic), and `PyTensor` that we use just to check if a value
is `torch.Tensor` and convert it into `tch::Tensor` in such a case.

The implementation of `PyTensor` is largerly taken from the `tch-pyo3` crate,
see the code for comments and the license.

What is not a part of this commit is to create `base::Term<>`
from `PyTerm`. It will be added in some later commits.