python: fix and improve torch integration
Use the new `Expr` class. It solves the problem that SymPy does not
make a difference between @ and *.
Also, now, if init or update is executed with a symbolic argument,
it is automagically traced instead of executed. If all arguments
are concrete, it executes normally. This enables the module
to be used arbitarily to create expressions.