Add `justfile`
A configuration to build the workspace in different setups.
It is useful, e.g., for building the python crate from scratch, because
that involves multiple steps. With `justfile`, we can make sure
that all these steps are done in order and e.g., all features
are set as they should.

Also update the README.